requirements.txt
<<<<<<< HEAD
# 문서 번역요약 서비스를 위한 파이썬 패키지들

# 웹 프레임워크
fastapi==0.104.1
uvicorn==0.24.0

# 번역 서비스 (GPT 기반)
openai==1.3.0

# 문서 처리
PyPDF2==3.0.1
python-docx==0.8.11

# 텍스트 요약 (경량화)
sumy==0.11.0

# 이미지 OCR
pytesseract==0.3.10
Pillow==10.1.0

# HTTP 요청
requests==2.31.0
httpx==0.25.2

# 환경 변수
python-dotenv==1.0.0

# JSON 처리
pydantic==2.5.0

# 파일 처리
aiofiles==23.2.1

# 로깅
loguru==0.7.2
=======
# 문서 번역요약 서비스를 위한 파이썬 패키지들

# 웹 프레임워크
fastapi==0.104.1
uvicorn==0.24.0

# 번역 서비스 (GPT 기반)
openai==1.3.0

# 문서 처리
PyPDF2==3.0.1
python-docx==0.8.11

# 텍스트 요약 (경량화)
sumy==0.11.0

# 이미지 OCR
pytesseract==0.3.10
Pillow==10.1.0

# HTTP 요청
requests==2.31.0
httpx==0.25.2

# 환경 변수
python-dotenv==1.0.0

# JSON 처리
pydantic==2.5.0

# 파일 처리
aiofiles==23.2.1

# 로깅
loguru==0.7.2
>>>>>>> 9472b9af39430ddd3198f479299d213279462014
